safeboda_rwanda/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.6.4
propcache==0.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
wheel==0.45.1
yarl==1.21.0
